refactor schema into its own file
